requirements.txt
aiosignal==1.3.1
appdirs==1.4.3
astroid==2.4.2
async-timeout==4.0.3; python_version < '3.11'
attrs==20.3.0
cachecontrol==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
discord.py==2.4.0
discord==1.0.1
distlib==0.3.0
distro==1.4.0
flask==1.1.2
frozenlist==1.4.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.7.0
itsdangerous==1.1.0
jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
markupsafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
multidict==5.1.0
pep517==0.8.2
progress==1.5
pycparser==2.20
pylint==2.6.0
pynacl==1.4.0
pyparsing==2.4.6
python-dotenv==1.0.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
websockets==8.1
werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
